+++ gtm/consentio-tag-cookies/src/sandbox.js
@@ -1,5 +1,0 @@
-const log = require('logToConsole');
-log('Consentio Tag: Cookies =', data);
-
-// Variables must return a value.
-return data.cookies;